+++ lib/ui/view/dm_chat_view/dm_jump_to_viewmodel.dart
@@ -4,5 +4,4 @@ import 'package:stacked/stacked.dart';
 class DmJumpToViewModel extends BaseViewModel {
   TextEditingController _controller = TextEditingController();
   TextEditingController get controller => _controller;
-
 }
